feat: add agent experience tracking
